Do not create PVC-bound deployments as part of main kustomization.yaml
This is because Rook w/ cephfs seems to get stuck provisioning them if
the Ceph cluster is not fully setup first. The 'volume already
in progress' error loop doesn't occur if the ceph cluster is setup and fully
provisioned before deploying PVCs and deployments which mount them.

This can be fixed by creating a Rook & Ceph cluster with a filesystem as part
of the ansible provisioning and waiting for it to be deployed.