Replace rook + ceph with longhorn + rook/NFS
This is due to cephFS's complexity. While rook and cephFS is very easy
to setup in a cluster for both block storage shared filesystem support,
it is very difficult to maintain and has issues with deploying it's
resources all at once. Also backup and recovery is not as straightforward
as longhorn.

It may be worth trying rook + cephFS once it easier to use and feels
less brittle. Rook with EdgeFS might be another option worth trying.

Longhorn is very easy to install and use, even if it is not
production ready. Combined with the Rook NFS operator, RWX is
possible by multiple nodes. This still does not allow for
horizontally scaling out the NFS operator across multiple
nodes due to the RWO block storage limitation compared to
cephFS CSI. However the easy remote and local backup are worth
the trade-off. It's significant simplicity made it a preference
over the more complicated Rook/cephFS.